+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.x branch
 ### 1.0.4-RC2
+* now it can load a configuration file;
 * removed `LinkScanner::setFullBaseUrl()` method. The full base url can be set
     as class configuration.